added Model to DTO converters, Update to Model, Update to DTOs

+++ app/app-core/src/main/java/iss/dt/app/core/model/Paper.java
@@ -14,8 +14,7 @@ import javax.persistence.*;
 public class Paper {
     @Id
     @GeneratedValue
-    private int id;
-
+    private Long id;
     private String name;
     private String abstractURL;
     private String fullURL;
